# src/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MS Producer API",
        "description": "API para cadastro de usuários com integração ao RabbitMQ",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor Local"
        }
    ],
    "paths": {
        "/api/users": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Listar todos os usuários",
                "description": "Retorna a lista de todos os usuários cadastrados no consumer",
                "operationId": "77c6da77a95ee5da609126d44ce70fe7",
                "responses": {
                    "200": {
                        "description": "Lista de usuários retornada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "string",
                                                        "format": "uuid",
                                                        "example": "9bcdcc8d-4bbb-42f0-8217-3c5b133ee6c9"
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "João da Silva"
                                                    },
                                                    "email": {
                                                        "type": "string",
                                                        "example": "joao.silva@example.com"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2025-11-28T03:06:11.000000Z"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2025-11-28T03:06:11.000000Z"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro ao buscar usuários",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Erro ao buscar usuários"
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Erro ao fazer requisição: Connection refused"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Cadastrar novo usuário",
                "description": "Cadastra um novo usuário e envia para processamento via RabbitMQ",
                "operationId": "fa44e1bcacaea5621fd93b0cdbd26a13",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Nome completo do usuário",
                                        "type": "string",
                                        "maxLength": 255,
                                        "minLength": 3,
                                        "example": "João da Silva"
                                    },
                                    "email": {
                                        "description": "E-mail do usuário",
                                        "type": "string",
                                        "format": "email",
                                        "example": "joao.silva@example.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Usuário cadastrado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Usuário enviado para processamento"
                                        },
                                        "data": {
                                            "properties": {
                                                "event_id": {
                                                    "type": "string",
                                                    "example": "user_656f8e4a5d1c83.12345678"
                                                },
                                                "user": {
                                                    "properties": {
                                                        "name": {
                                                            "type": "string",
                                                            "example": "João da Silva"
                                                        },
                                                        "email": {
                                                            "type": "string",
                                                            "example": "joao.silva@example.com"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erro de validação",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "O nome deve ter no mínimo 3 caracteres (and 1 more error)"
                                        },
                                        "errors": {
                                            "properties": {
                                                "name": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "O nome deve ter no mínimo 3 caracteres"
                                                    }
                                                },
                                                "email": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "O e-mail deve ser válido"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno do servidor",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Erro ao processar requisição"
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Não foi possível conectar ao RabbitMQ"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/users/{id}": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Buscar usuário específico",
                "description": "Retorna os dados de um usuário específico pelo ID",
                "operationId": "bfd755475b7a7865c3d1745f2e483eed",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do usuário (UUID)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "example": "9bcdcc8d-4bbb-42f0-8217-3c5b133ee6c9"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Usuário encontrado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "string",
                                                    "format": "uuid",
                                                    "example": "9bcdcc8d-4bbb-42f0-8217-3c5b133ee6c9"
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "João da Silva"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "example": "joao.silva@example.com"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2025-11-28T03:06:11.000000Z"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2025-11-28T03:06:11.000000Z"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Usuário não encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Usuário não encontrado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro ao buscar usuário",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Erro ao buscar usuário"
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Erro ao fazer requisição: Connection refused"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/health": {
            "get": {
                "tags": [
                    "Health"
                ],
                "summary": "Health check",
                "description": "Verifica se a API está rodando corretamente",
                "operationId": "df8311a6d3330593a9f6287a3c36bc3a",
                "responses": {
                    "200": {
                        "description": "API está funcionando",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "service": {
                                            "type": "string",
                                            "example": "ms-producer"
                                        },
                                        "status": {
                                            "type": "string",
                                            "example": "running"
                                        },
                                        "timestamp": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2025-11-28T03:06:11+00:00"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Health",
            "description": "Endpoints de health check"
        },
        {
            "name": "Users",
            "description": "Endpoints de gerenciamento de usuários"
        }
    ]
}